feat(group): add shared group creation and refactor home group restoration logic with tests (#179)
