+++ src/test/java/com/solv/wefin/domain/group/service/GroupJoinServiceTest.java
@@ -1,0 +1,431 @@
+package com.solv.wefin.domain.group.service;
+
+import com.solv.wefin.domain.auth.repository.UserRepository;
+import com.solv.wefin.domain.group.dto.GroupMemberInfo;
+import com.solv.wefin.domain.group.entity.Group;
+import com.solv.wefin.domain.group.entity.GroupInvite;
+import com.solv.wefin.domain.group.entity.GroupMember;
+import com.solv.wefin.domain.group.entity.GroupType;
+import com.solv.wefin.domain.group.repository.GroupInviteRepository;
+import com.solv.wefin.domain.group.repository.GroupMemberRepository;
+import com.solv.wefin.domain.group.repository.GroupRepository;
+import com.solv.wefin.global.error.BusinessException;
+import com.solv.wefin.global.error.ErrorCode;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.time.OffsetDateTime;
+import java.util.Optional;
+import java.util.UUID;
+
+import static com.solv.wefin.domain.group.service.support.GroupTestFixtures.createGroup;
+import static com.solv.wefin.domain.group.service.support.GroupTestFixtures.createGroupInvite;
+import static com.solv.wefin.domain.group.service.support.GroupTestFixtures.createGroupMember;
+import static com.solv.wefin.domain.group.service.support.GroupTestFixtures.createUser;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.anyLong;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+@ExtendWith(MockitoExtension.class)
+class GroupJoinServiceTest {
+
+    @Mock
+    private GroupRepository groupRepository;
+
+    @Mock
+    private GroupMemberRepository groupMemberRepository;
+
+    @Mock
+    private GroupInviteRepository groupInviteRepository;
+
+    @Mock
+    private UserRepository userRepository;
+
+    @InjectMocks
+    private GroupService groupService;
+
+    @Nested
+    @DisplayName("joinGroup")
+    class JoinGroupTest {
+
+        @Test
+        @DisplayName("초대 코드로 SHARED 그룹에 정상 참여한다")
+        void joinGroup_success() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            Group homeGroup = createGroup(100L, "홈 그룹", GroupType.HOME);
+
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            GroupMember currentActiveHomeMember = createGroupMember(
+                    20L,
+                    user,
+                    homeGroup,
+                    GroupMember.GroupMemberRole.LEADER,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            );
+
+            GroupMember newMember = createGroupMember(
+                    21L,
+                    user,
+                    targetGroup,
+                    GroupMember.GroupMemberRole.MEMBER,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            );
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+            when(groupRepository.findByIdForUpdate(targetGroup.getId()))
+                    .thenReturn(Optional.of(targetGroup));
+            when(userRepository.findById(userId))
+                    .thenReturn(Optional.of(user));
+            when(groupMemberRepository.findByUser_UserIdAndStatus(
+                    userId,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(Optional.of(currentActiveHomeMember));
+            when(groupMemberRepository.countByGroupAndStatus(
+                    targetGroup,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(1L);
+            when(groupMemberRepository.findByUser_UserIdAndGroup_Id(userId, targetGroup.getId()))
+                    .thenReturn(Optional.empty());
+            when(groupMemberRepository.save(any(GroupMember.class)))
+                    .thenReturn(newMember);
+
+            // when
+            GroupMemberInfo result = groupService.joinGroup(userId, inviteCode);
+
+            // then
+            assertAll(
+                    () -> assertThat(result.groupId()).isEqualTo(1L),
+                    () -> assertThat(result.groupName()).isEqualTo("공유 그룹"),
+                    () -> assertThat(result.role()).isEqualTo("MEMBER"),
+                    () -> assertThat(currentActiveHomeMember.isActive()).isFalse(),
+                    () -> assertThat(invite.getStatus()).isEqualTo(GroupInvite.InviteStatus.ACCEPTED)
+            );
+
+            verify(groupMemberRepository).save(any(GroupMember.class));
+        }
+
+        @Test
+        @DisplayName("기존 멤버십이 있으면 재활성화한다")
+        void joinGroup_success_when_membership_exists() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            Group homeGroup = createGroup(100L, "홈 그룹", GroupType.HOME);
+
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            GroupMember currentActiveHomeMember = createGroupMember(
+                    20L,
+                    user,
+                    homeGroup,
+                    GroupMember.GroupMemberRole.LEADER,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            );
+
+            GroupMember existingMembership = createGroupMember(
+                    21L,
+                    user,
+                    targetGroup,
+                    GroupMember.GroupMemberRole.MEMBER,
+                    GroupMember.GroupMemberStatus.INACTIVE
+            );
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+            when(groupRepository.findByIdForUpdate(targetGroup.getId()))
+                    .thenReturn(Optional.of(targetGroup));
+            when(userRepository.findById(userId))
+                    .thenReturn(Optional.of(user));
+            when(groupMemberRepository.findByUser_UserIdAndStatus(
+                    userId,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(Optional.of(currentActiveHomeMember));
+            when(groupMemberRepository.countByGroupAndStatus(
+                    targetGroup,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(1L);
+            when(groupMemberRepository.findByUser_UserIdAndGroup_Id(userId, targetGroup.getId()))
+                    .thenReturn(Optional.of(existingMembership));
+
+            // when
+            GroupMemberInfo result = groupService.joinGroup(userId, inviteCode);
+
+            // then
+            assertAll(
+                    () -> assertThat(result.groupId()).isEqualTo(1L),
+                    () -> assertThat(result.groupName()).isEqualTo("공유 그룹"),
+                    () -> assertThat(result.role()).isEqualTo("MEMBER"),
+                    () -> assertThat(currentActiveHomeMember.isActive()).isFalse(),
+                    () -> assertThat(existingMembership.isActive()).isTrue(),
+                    () -> assertThat(invite.getStatus()).isEqualTo(GroupInvite.InviteStatus.ACCEPTED)
+            );
+
+            verify(groupMemberRepository, never()).save(any(GroupMember.class));
+        }
+
+        @Test
+        @DisplayName("홈 그룹에는 참여할 수 없다")
+        void joinGroup_fail_when_home_group() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group homeGroup = createGroup(1L, "홈 그룹", GroupType.HOME);
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    homeGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+            when(groupRepository.findByIdForUpdate(homeGroup.getId()))
+                    .thenReturn(Optional.of(homeGroup));
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_HOME_JOIN_NOT_ALLOWED);
+
+            verify(userRepository, never()).findById(any());
+            verify(groupMemberRepository, never()).findByUser_UserIdAndStatus(any(), any());
+            verify(groupMemberRepository, never()).countByGroupAndStatus(any(), any());
+            verify(groupMemberRepository, never()).findByUser_UserIdAndGroup_Id(any(), anyLong());
+            verify(groupMemberRepository, never()).save(any());
+        }
+
+        @Test
+        @DisplayName("이미 같은 그룹에 ACTIVE 상태로 참여 중이면 예외가 발생한다")
+        void joinGroup_fail_when_already_joined() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            GroupMember currentActiveMember = createGroupMember(
+                    20L,
+                    user,
+                    targetGroup,
+                    GroupMember.GroupMemberRole.MEMBER,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            );
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+            when(groupRepository.findByIdForUpdate(targetGroup.getId()))
+                    .thenReturn(Optional.of(targetGroup));
+            when(userRepository.findById(userId))
+                    .thenReturn(Optional.of(user));
+            when(groupMemberRepository.findByUser_UserIdAndStatus(
+                    userId,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(Optional.of(currentActiveMember));
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_ALREADY_JOINED);
+
+            verify(groupMemberRepository, never()).countByGroupAndStatus(any(), any());
+            verify(groupMemberRepository, never()).findByUser_UserIdAndGroup_Id(any(), anyLong());
+            verify(groupMemberRepository, never()).save(any());
+        }
+
+        @Test
+        @DisplayName("정원이 가득 찬 그룹은 참여할 수 없다")
+        void joinGroup_fail_when_group_full() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            Group homeGroup = createGroup(100L, "홈 그룹", GroupType.HOME);
+
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            GroupMember currentActiveHomeMember = createGroupMember(
+                    20L,
+                    user,
+                    homeGroup,
+                    GroupMember.GroupMemberRole.LEADER,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            );
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+            when(groupRepository.findByIdForUpdate(targetGroup.getId()))
+                    .thenReturn(Optional.of(targetGroup));
+            when(userRepository.findById(userId))
+                    .thenReturn(Optional.of(user));
+            when(groupMemberRepository.findByUser_UserIdAndStatus(
+                    userId,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(Optional.of(currentActiveHomeMember));
+            when(groupMemberRepository.countByGroupAndStatus(
+                    targetGroup,
+                    GroupMember.GroupMemberStatus.ACTIVE
+            )).thenReturn(6L);
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_FULL);
+
+            verify(groupMemberRepository, never()).findByUser_UserIdAndGroup_Id(any(), anyLong());
+            verify(groupMemberRepository, never()).save(any());
+        }
+
+        @Test
+        @DisplayName("만료된 초대 코드는 사용할 수 없다")
+        void joinGroup_fail_when_invite_expired() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+
+            invite.expire();
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_INVITE_EXPIRED);
+        }
+
+        @Test
+        @DisplayName("이미 사용된 초대 코드는 사용할 수 없다")
+        void joinGroup_fail_when_invite_already_used() throws Exception {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            Group targetGroup = createGroup(1L, "공유 그룹", GroupType.SHARED);
+            var user = createUser(userId, "test@test.com", "유저", "pw");
+
+            GroupInvite invite = createGroupInvite(
+                    10L,
+                    targetGroup,
+                    user,
+                    inviteCode,
+                    GroupInvite.InviteStatus.PENDING
+            );
+            invite.markAccepted();
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.of(invite));
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_INVITE_ALREADY_USED);
+        }
+
+        @Test
+        @DisplayName("존재하지 않는 초대 코드는 예외가 발생한다")
+        void joinGroup_fail_when_invite_not_found() {
+            // given
+            UUID userId = UUID.randomUUID();
+            UUID inviteCode = UUID.randomUUID();
+
+            when(groupInviteRepository.findByInviteCode(inviteCode))
+                    .thenReturn(Optional.empty());
+
+            // when
+            BusinessException exception = assertThrows(
+                    BusinessException.class,
+                    () -> groupService.joinGroup(userId, inviteCode)
+            );
+
+            // then
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.GROUP_INVITE_NOT_FOUND);
+        }
+    }
+}
